requirements.txt
unstructured
unstructured[pdf,docx,image]
unstructured-inference
pillow-heif
pdf2image
python-dotenv
pinecone
openai
-e .